Splitting getOptions across data types
